Errores ortográficos, 0 es considerada una tempertura positiva, la variable i es entera, es decir, int

+++ ejercicio-practico/src/ejercicio/practico/EjercicioPractico.java
@@ -18,7 +18,7 @@ public class EjercicioPractico {
         System.out.print("Introduce el numero de temperaturas que desea introducir:\n");
         int n;
         n = entrada.nextInt();
-        for (float i = 1; i <= n; i++) {
+        for (int i = 1; i <= n; i++) {
 
             System.out.print("Introduce una temperatura:\n");
             temp = entrada.nextFloat();
